+++ EntraAuth/internal/classes/Environment.ps1
@@ -1,0 +1,6 @@
+﻿enum Environment {
+	Global = 1
+	USGov = 2
+	USGovDOD = 3
+	China = 4
+}