+++ EntraAuth/changelog.md
@@ -1,5 +1,11 @@
 ﻿# Changelog
 
+## 1.3.18 (2024-10-08)
+
++ Upd: Added support for authenticating using the current Az.Accounts session
++ Upd: Added support for Sovereign Clouds (USGov, USGovDOD, China) and custom authentication urls.
++ Fix: Certificate Logon fails on timezones after UTC
+
 ## 1.2.15 (2024-07-31)
 
 + Fix: Refresh token may fail to authenticate to correct application